Add basic demo script
The script clones the demo jekyll project and runs the
supply chain commands as defined in the layout using in-toto
and eventually verifies the supply chain (and passes).

There is an optional "attack" argument that runs the same comamnds
but inserts malicious code in between the in-toto steps, which
makes verification fail.