Add basic grafeas demo script
Pushes the layout to a grafeas server (as operation metatada).

Then runs the supply chain commands using the grafeas in-toto client,
that in addition to creating in-toto metadata, makes the metadata
greafeas compliant and pushes it to a grafeas server (currently
the default server hosted by Santiago).

Eventually the script fetches all the metadata from the server
and performs (a passing) in-toto verification.